+++ app/src/main/java/com/example/filmly/MainActivity.kt
@@ -1,6 +1,5 @@
-package com.digitalhouse.projetointegrador
+package com.example.filmly
 
-import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.os.Handler
