+++ app/src/main/java/com/example/filmly/SplashActivity.kt
@@ -1,4 +1,4 @@
-package com.digitalhouse.projetointegrador
+package com.example.filmly
 
 import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
