+++ app/src/main/java/com/example/filmly/ViewMoreFragment.kt
@@ -1,14 +1,13 @@
-package com.digitalhouse.projetointegrador
+package com.example.filmly
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.view.animation.AnimationUtils
 import androidx.recyclerview.widget.GridLayoutManager
 import androidx.recyclerview.widget.RecyclerView
-import com.digitalhouse.projetointegrador.domain.ViewMoreCard
+import com.example.filmly.domain.ViewMoreCard
 
 
 class ViewMoreFragment : Fragment() {
