+++ app/src/test/java/com/example/filmly/ExampleUnitTest.kt
@@ -1,4 +1,4 @@
-package com.digitalhouse.projetointegrador
+package com.example.filmly
 
 import org.junit.Test
